node-editor: implemented cursor for text box widget

+++ node-editor/include/Widget.h
@@ -17,7 +17,7 @@ namespace Widgets
     struct TextBoxState
     {
         std::string data{};
-        int cursor_position = 0;
+        size_t cursor_position = 0;
     };
 
     struct None
